Enter the gallery, soundrecorder and other applications, Select file delete and press the allow button. Because "allow button" is clicked continuously, resulting in the repeated creation of stack top activities. The dialog box requesting deletion appears many times, but the file has been deleted at this time, which will cause misunderstanding.
When test the monkey, continuous clicking on "delete button" and the
default startup mode of the system is standard will start several
activities, resulting in repeated creation of stack top activities.

In order to set successfully after clicking allow for the first time,
there are no other exceptions, Set the startup mode of the activity to
singletask in Android manifest.xml

Test: Perform the monkey test, and see if Select the files to delete,
The dialog box appears and then click the allow button to delete the
Selected files  successfully. The dialog will happen again when the
default startup mode is standard.

Bug: 213043426
Change-Id: Ia7f217ea180961321e0a9108209907e5100ec229
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>