Fixed deadlock issue caused by fuse_inval call in invalidateFuseDentry
When an application uses the file interface to rename a file to a directory, applyBatch of the media library is called to rename the files in the directory (for example,. Trash). Deadlock occurs. Specific process:
1, Rename file holds lock A in kernel open, sleep, call back to media library insertFileForFuse
When applyBatch is called, the transaction is started and the update interface is called. Fuse_inval is called inside the update interface. Fuse_inval is waiting for lock A
InsertFileForFuse is waiting to start a transaction, but 2. The transaction in is not finished because the internal fuse_inval is waiting for lock A.

Open a new thread to call fuse_inval, allowing the update interface to end the transaction without causing a deadlock.

Bug: 238933498
Change-Id: If9284a3e60747353805fd6e2dc6512259c7eada0
Signed-off-by: strwbrry143 <alexfinhart@gmail.com>